Add connectors for rows and columns
Changes have been made to add connectors for the rows and columns so that two PCBs can be bridged together to create a single 12x5 matrix running from a single Pro Micro.
Provision for removal of the bottom row has been removed to make way for the connectors.
Holes for 2U stablizers have been removed to increase PCB stiffness.